refactor(전체): domain과 view의 분리
