+++ src/main/java/step3/domain/Car.java
@@ -1,8 +1,7 @@
-package step3;
+package step3.domain;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Random;
 
 public class Car {
