+++ src/main/java/step3/domain/Round.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.domain;
 
 public class Round {
     private int time;
