+++ src/main/java/step3/utility/ValidationUtils.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.utility;
 
 public class ValidationUtils {
