+++ src/test/java/step3/CarTest.java
@@ -3,9 +3,7 @@ package step3;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import java.util.HashSet;
-import java.util.Set;
+import step3.domain.Car;
 
 import static org.assertj.core.api.Assertions.assertThat;
